Twitter Search scripts working with posrgres
The search script are initally working. However the commit operation needs to
be only called at the end of a page, for performance. The page is returning
duplicate tweet id's, this has been fixed by surrounding the query with
try/catch but maybe there is a better solution.

Finally every time this version of the script runs, it DROPS the current db
table. This is needed for debug... you have been warned......