Stop storing retweets & get search terms from file
Every tweet is checked for the string "RT @", if this string is found assume to
be a retweet and drop the message. This does not cover the case where a retweet
is stored in metadata - as is now much more common.

Search terms are now externalised to a text file. This will be easier to track.