Added script to create new clusters for time slice
In this implementation kmeans is used and Kullback-Leibler divergence.
At this state it prints out the values of KL divergence for the new
clusters comparing them to the static clusters.
Because geolondon doesn't have new data and usage of all tweets was
required and not only for traffic, the datetime to which we're looking at
tweets is the last datetime anything was added to this table (2012/03/03,
2:17:18)